+++ trm/Throat.go
@@ -28,15 +28,15 @@
 package trm
 
 type Throat struct {
-	Gain float64
-	tb1  float64
-	ta0  float64
-	Y    float64
+	Gain float32
+	tb1  float32
+	ta0  float32
+	Y    float32
 }
 
 // Init Initializes the throat lowpass filter coefficients according to the throatCutoff value,
 // and also the throatGain, according to the throatVol value.
-func (thr *Throat) Init(sampleRate, cutoff, gain float64) {
+func (thr *Throat) Init(sampleRate, cutoff, gain float32) {
 	thr.Gain = gain
 	thr.ta0 = (cutoff * 2.0) / sampleRate
 	thr.tb1 = 1.0 - thr.ta0
@@ -49,7 +49,7 @@ func (thr *Throat) Reset() {
 
 // Process Simulates the radiation of sound through the walls of the throat.
 // Note that this form of the filter uses addition instead of subtraction for the econd term, since tb1 has reversed sign.
-func (thr *Throat) Process(input float64) float64 {
+func (thr *Throat) Process(input float32) float32 {
 	output := (thr.ta0 * input) + (thr.tb1 * thr.Y)
 	thr.Y = output
 	return output * thr.Gain
